fix issue #1 by waiting for runway to clear
https://github.com/couchbaselabs/blance/issues/1

This change has the supreme airport controller (supplier of moves)
wait for a runMover to fully process a move before supplying another
move to a runMover.

In contrast, the previous code just had the supreme controller just
wait for a runMover to accept a move, but incorrectly not wait for
completion of that move request.